# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CopyPasteHouse API",
        "description": "API documentation for CopyPasteHouse - A paste sharing application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/pastes": {
            "get": {
                "tags": [
                    "Pastes"
                ],
                "summary": "List all pastes",
                "description": "Retrieve a paginated list of pastes with optional filters",
                "operationId": "24b693a2e24b6551837d2d7ff5b2c721",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of pastes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    },
                                                    "content": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Pastes"
                ],
                "summary": "Create a new paste",
                "description": "Create a new paste with title and content",
                "operationId": "caef8867e233cd292a34e4f09a3734cf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "My Code Snippet"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "console.log(\"Hello World\")"
                                    },
                                    "syntax_highlight_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "expiration_time_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "password": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Paste created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "paste": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Paste created successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/pastes/my-pastes": {
            "get": {
                "tags": [
                    "Pastes"
                ],
                "summary": "List authenticated user pastes",
                "description": "Retrieve a list of pastes created by the authenticated user",
                "operationId": "0c933e76106618965382b2c29e51bad5",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of user pastes"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/pastes/{paste}": {
            "get": {
                "tags": [
                    "Pastes"
                ],
                "summary": "Get a specific paste",
                "description": "Retrieve details of a specific paste by ID",
                "operationId": "45d3c037bbcc1970ae6cfb7059557110",
                "parameters": [
                    {
                        "name": "paste",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paste details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "content": {
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Paste not found"
                    },
                    "403": {
                        "description": "Password required"
                    }
                }
            },
            "put": {
                "tags": [
                    "Pastes"
                ],
                "summary": "Update a paste",
                "description": "Update an existing paste (requires authentication)",
                "operationId": "03f7d35819b788dda21c3f4531e54a02",
                "parameters": [
                    {
                        "name": "paste",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "syntax_highlight_id": {
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paste updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "paste": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Paste updated successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Paste not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pastes"
                ],
                "summary": "Delete a paste",
                "description": "Delete an existing paste (requires authentication)",
                "operationId": "fb61f1f4f9050d835337b3df7d04a268",
                "parameters": [
                    {
                        "name": "paste",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paste deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Paste deleted successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Paste not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/pastes/{paste}/like": {
            "post": {
                "tags": [
                    "Pastes"
                ],
                "summary": "Toggle like on paste",
                "description": "Like or unlike a paste (requires authentication)",
                "operationId": "70a7c39c86b5eda33c13620e0070df50",
                "parameters": [
                    {
                        "name": "paste",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Like toggled successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Paste liked!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Paste not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}